requirements.txt
alembic==1.12.1
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
click==8.1.7
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
flask-restx==1.2.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyrsistent==0.20.0
python-decouple==3.8
pytz==2023.3.post1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==2.3.8
